Implementation Iteration 1 (v 0.1) – Basic Setup
In my first iteration, I just mainly set pygame up with the game loop and made my main menu scene. I need this game loop so that the game does not just end after one pass and all events that the player triggers (such as mouse clicks, button presses etc.) are registered through it. It also allows me to pass arguments (namely events) to my scenes which will handle them separately based on the scene we are on.
One may also notice that I have set up the barebones of some classes. This was done so that I have a general overview of what else I need to develop and so that I can do parts in parallel and think about the details later on.
Currently, as what I have are less than the barebones, there is no validation. All I have is a static, non-interactive picture.