Between Iteration
I have managed to make the hovering behaviour to be detected, however now when I hover over a button, the colour which the button gets changed to negative values in the colour tuple, making it invalid and crashing my program. I need to look into a fix for this.

This was a logical error; each time I entered the method, the onceOver variable would renew its status without affecting anything. All I had to do was make it an instance variable with a default value of zero and remove the line in the method which sets its value to zero at the start.