v 0.3 - working on graphics
(Note: the 0.2.x versions might not be in the correct order, however this comes after all of them)

In this iteration, I have completely finished and polished the graphical aspect of my game. This means that on the surface, my game looks complete, and if someone was not paying much attention to someone interacting with the game, they might say it is finished. I have completed the GameScene class which contains all of the elements a user is going to be interacting with when they are actually playing the game (this involves buttons the board, their rack etc.). The reason I did this is so that the user can have a good experience in playing the game, and instead of doing everything via a console, they can do it with a GUI. It also allows me to now focus on just the logic of the game without having to worry at all about how it is going to affect the looks of the game, as I have completely separated the logical and visual part of my game.

Firstly, let’s start from the top: the Director received a minor change, where instead of checking if a certain selection has been made to trigger the ‘NotImplementedScene’ (see line 55 in previous iteration of Director), whenever an error "Section of game not implemented yet" is raised, the ‘NotImplementedScene’ event is triggered. Also, as pygame implements its own wait procedure, I no longer require the time module, and have changed that one line to reflect this realisation.

I have yet to implement validation checking for words. This is likely to come later on in my iterations, as it requires checking of validity of words, which is most easily done with GADDAG. It will also be handled through the Game class, so this class is finished with respect to that.