Implementation Iteration 4 (v 1.0/0.4) – finishing player vs. player
For this iteration, I firstly looked at my user feedback and decided to make improvements to my program based on what my user wanted. This was to display which tile they have selected in some meaningful way; this was only a minor change, but makes the program better, as the user gets feedback based on their actions.
I then moved onto the most important part of the game: checking whether the user’s input is actually valid. It required me to check that all potential words submitted in this turn are valid; this problem breaks down into firstly making sure that the user places either only one tile, places tiles down only in one row or only in one column. Next, I had to make sure that these tiles are connected to the centre tile, either directly, or through other tiles. And lastly, I had to make sure that all of the words which were submitted were actually valid words (i.e. in my dictionary file). If the user’s input passes all of this validation, then their input is valid, and they successfully submit their word. Otherwise, after pressing submit word, nothing happens, which signifies that one of these validation stages has failed, according to the rules of Words With Friends.
The next and last thing (other than some minor bug fixes/improvements) that I needed to do to finish my game was to make sure that once a valid word has been submitted, the player who submits it gets credited the correct amount of points for the word they submitted, again just following the rules (and my interpretation of said rules) of Words With Friends (found on https://www.zyngawithfriends.com/wordswithfriends/support/WWF_Rulebook.html on 29.01.2019).

Because checking for winner has to be done at the end of each turn, and my code checks for such changes at the start of each turn (BeforeEachTurn()), I decided to rework this method into a AfterEachTurn() method, as otherwise the if there is a winner, they would be announced after the players change turns (after the 5 seconds pause) which doesn’t look nice. This required changes in the PassTurn() method.

While testing my program out, I also realised that if I select a tile, and then shuffle my tiles, the selection persists on the same tile, which is shown by the tile being deactivated. This was unwanted, and so to fix it, I added a call to the CancelMoves() method inside of the ShuffleTiles() method

There were occasions on which my game would crash, and I wasn’t sure why, so I decided to investigate this. The problem turned out to be that a TextBox object somehow ended up with a GLetterTile as its text attribute, and when narrowing down on what causes this, I found out that my WildCardPicker ProcessInput() method  returns a GLetterTile instead of a string, so I fixed this

Another oversight I found out while testing was that if the ‘Tile Bag’ button is pressed when there aren’t instances of every tiles passed into its constructor, it either crashes or the wrong numbers are displayed. This was because I relied on every letter being passed in. To fix this, instead of self.tiles and self.numbers_of_tiles being lists, I made then dictionaries, with the letters as keys, and the tiles and numbers of tiles as the values, respectively. This allowed me to display zero instances of a tile correctly.
And to accommodate for this change, I also needed to change the SetPosition(), Draw() and ProcessInput() methods so that instead of iterating through a list (or enumerated list), I iterated through dictionaries.