Implementation Iteration 5 (v 1.1) – finishing player vs. player
In this iteration, I decided to just improve the minor things which I got as feedback from my prototype testing. These were the ‘making it obvious who the current player is’ and ‘invalid word message’. The reason for doing this is to satisfy the needs of the user.

As in this iteration I have not added extra functionality to the game which the user has control over, I have had no need to consider any validation, as what the user can interact with has stayed the same; they’re just receiving more feedback.