Implementation Iteration 7 (v 1.2) - AI development
The main goal of this iteration was to fully implement an AI into my game, and it turned out successful.
To achieve this, I had to make use of almost every logical part of my game which I had coded, and some of the previously coded parts required a bit of rethinking (such as how does the AI know when it is its turn, or how are its tiles actually displayed onto the visual board, which was partly done before in one of the iterations).

there was an oversight by me for when the user clicks on a tile, and then clicks on the shuffle tiles button which made it so that the clicked tiles would persist through the shuffling or worse, word swapping, and would not be interactive. To fix this, I included an extra line at the end of ReturnMovedTilesToRack().