Implementation Iteration 8 (v 1.2.1) - fixing up the AI
From user testing, I received feedback that my AI does not consider the situation when it can add an extra letter to the end of a word, and then form a word in the other axis. For example, if the word ‘hair’ was in the horizontal axis, and the AI had the letters ‘s’, ‘e’, and ‘t’, it would not be able to form the word ‘set’ in the vertical axis. It would instead play the word ‘hairs’. In this iteration, I addressed and fixed this issue by changing a few things in HookH() and HookV().

So, instead of just saying whether there was a tile placed down, I count how many tiles were placed down. If only one was placed down, then there is a possibility that the above mentioned case could have occurred. There is also the extra parameter of ‘v_tried’ and ‘h_tried’ which just make sure that no infinite recursion occurs, as if I tried to hook onto that one newly placed tile with the different axis hook, then it might try to hook back onto where it came from, causing infinite recursion. The ‘x_new’ and ‘y_new’ are added in so that I can also have access to the previous x and y values, as these are the ones I would need to hook onto if the are wanting to hook on the other axis.
So, the way it is done, is the addition of line 225 and 226 (and respectively in the other axis method), which check the above, and also that the formed two letter word is actually a valid word. If so, we can try hooking onto the one newly placed lettertile in the other axis, and forming a word there.